+++ app/models/Classroom.php
@@ -167,6 +167,7 @@ class Classroom extends AltActiveRecord
             'calendarFk' => array(self::BELONGS_TO, 'Calendar', 'calendar_fk'),
             'schoolInepFk' => array(self::BELONGS_TO, 'SchoolIdentification', 'school_inep_fk'),
             'edcensoStageVsModalityFk' => array(self::BELONGS_TO, 'EdcensoStageVsModality', 'edcenso_stage_vs_modality_fk'),
+            'classroomHasCoursePlans' => array(self::HAS_MANY, 'ClassroomHasCoursePlan', 'classroom_fk'),
             'calendarFk' => array(self::BELONGS_TO, 'Calendar', 'calendar_fk'),
             'sedspSchoolUnityFk' => array(self::BELONGS_TO, 'SedspSchoolUnities', 'sedsp_school_unity_fk'),
             'instructorTeachingDatas' => array(self::HAS_MANY, 'InstructorTeachingData', 'classroom_id_fk'),
@@ -195,7 +196,7 @@ class Classroom extends AltActiveRecord
                 self::STAT,
                 'StudentEnrollment',
                 'classroom_fk',
-                'condition' => 'status IN (1, 6, 7, 8, 9, 10) or status IS NULL'
+                'condition' => 'status IN (1, 2, 6, 7, 8, 9, 10) or status IS NULL'
             ),
         );
     }
@@ -290,6 +291,22 @@ class Classroom extends AltActiveRecord
             'sedsp_max_physical_capacity' => "Sedsp Max Physical Capacity",
             'gov_id' => "GOV ID"
         );
+    }
+
+
+    public function getGradeRules($stageId=null)   {
+        if ($stageId == null){
+            $stageId = $this->edcenso_stage_vs_modality_fk;
+        }
+
+        $criteria = new CDbCriteria();
+        $criteria->alias = 'gr';
+        $criteria->join = 'join grade_rules_vs_edcenso_stage_vs_modality grvesvm on gr.id = grvesvm.grade_rules_fk';
+        $criteria->join .= ' join classroom_vs_grade_rules cvgr on cvgr.grade_rules_fk = gr.id';
+        $criteria->condition = 'grvesvm.edcenso_stage_vs_modality_fk = :stage and cvgr.classroom_fk = :classroom';
+        $criteria->params = array(':classroom' => $this->id, ':stage' => $stageId);
+        return GradeRules::model()->find($criteria);
+
     }
 
     /**