+++ app/models/ClassroomVsGradeRules.php
@@ -1,31 +1,29 @@
 <?php
 
 /**
- * This is the model class for table "grade_rules".
+ * This is the model class for table "classroom_vs_grade_rules".
  *
- * The followings are the available columns in table 'grade_rules':
+ * The followings are the available columns in table 'classroom_vs_grade_rules':
  * @property integer $id
- * @property integer $edcenso_stage_vs_modality_fk
- * @property double $approvation_media
- * @property double $final_recover_media
- * @property integer $grade_calculation_fk
- * @property integer $has_final_recovery
- * @property integer $has_partial_recovery
- * @property string $rule_type
+ * @property integer $classroom_fk
+ * @property integer $grade_rules_fk
+ * @property string $created_at
+ * @property string $updated_at
  *
  * The followings are the available model relations:
- * @property GradeCalculation $gradeCalculationFk
- * @property EdcensoStageVsModality $edcensoStageVsModalityFk
+ * @property Classroom $classroomFk
+ * @property GradeRules $gradeRulesFk
  */
-class GradeRules extends CActiveRecord
+class ClassroomVsGradeRules extends CActiveRecord
 {
 	/**
 	 * @return string the associated database table name
 	 */
 	public function tableName()
 	{
-		return 'grade_rules';
+		return 'classroom_vs_grade_rules';
 	}
+
     public function behaviors()
     {
         return [
@@ -39,6 +37,7 @@ class GradeRules extends CActiveRecord
         ];
     }
 
+
 	/**
 	 * @return array validation rules for model attributes.
 	 */
@@ -47,13 +46,11 @@ class GradeRules extends CActiveRecord
 		// NOTE: you should only define rules for those attributes that
 		// will receive user inputs.
 		return array(
-			array('approvation_media', 'required', 'on' => "numericGrade"),
-			array('grade_calculation_fk, has_final_recovery, has_partial_recovery', 'numerical', 'integerOnly'=>true),
-			array('approvation_media, final_recover_media', 'numerical'),
-			array('rule_type', 'length', 'max'=>1),
+			array('classroom_fk, grade_rules_fk', 'required'),
+			array('classroom_fk, grade_rules_fk', 'numerical', 'integerOnly'=>true),
 			// The following rule is used by search().
 			// @todo Please remove those attributes that should not be searched.
-			array('id, edcenso_stage_vs_modality_fk, approvation_media, final_recover_media, grade_calculation_fk, has_final_recovery, has_partial_recovery, rule_type', 'safe', 'on'=>'search'),
+			array('id, classroom_fk, grade_rules_fk, created_at, updated_at', 'safe', 'on'=>'search'),
 		);
 	}
 
@@ -65,8 +62,8 @@ class GradeRules extends CActiveRecord
 		// NOTE: you may need to adjust the relation name and the related
 		// class name for the relations automatically generated below.
 		return array(
-			'gradeCalculationFk' => array(self::BELONGS_TO, 'GradeCalculation', 'grade_calculation_fk'),
-			'edcensoStageVsModalityFk' => array(self::BELONGS_TO, 'EdcensoStageVsModality', 'edcenso_stage_vs_modality_fk'),
+			'classroomFk' => array(self::BELONGS_TO, 'Classroom', 'classroom_fk'),
+			'gradeRulesFk' => array(self::BELONGS_TO, 'GradeRules', 'grade_rules_fk'),
 		);
 	}
 
@@ -77,13 +74,10 @@ class GradeRules extends CActiveRecord
 	{
 		return array(
 			'id' => 'ID',
-			'edcenso_stage_vs_modality_fk' => 'Edcenso Stage Vs Modality Fk',
-			'approvation_media' => 'Approvation Media',
-			'final_recover_media' => 'Final Recover Media',
-			'grade_calculation_fk' => 'Grade Calculation Fk',
-			'has_final_recovery' => 'Has Final Recovery',
-			'has_partial_recovery' => 'Has Partial Recovery',
-			'rule_type' => 'Rule Type',
+			'classroom_fk' => 'Classroom Fk',
+			'grade_rules_fk' => 'Grade Rules Fk',
+			'created_at' => 'Created At',
+			'updated_at' => 'Updated At',
 		);
 	}
 
@@ -106,13 +100,10 @@ class GradeRules extends CActiveRecord
 		$criteria=new CDbCriteria;
 
 		$criteria->compare('id',$this->id);
-		$criteria->compare('edcenso_stage_vs_modality_fk',$this->edcenso_stage_vs_modality_fk);
-		$criteria->compare('approvation_media',$this->approvation_media);
-		$criteria->compare('final_recover_media',$this->final_recover_media);
-		$criteria->compare('grade_calculation_fk',$this->grade_calculation_fk);
-		$criteria->compare('has_final_recovery',$this->has_final_recovery);
-		$criteria->compare('has_partil_recovery',$this->has_partial_recovery);
-		$criteria->compare('rule_type',$this->rule_type,true);
+		$criteria->compare('classroom_fk',$this->classroom_fk);
+		$criteria->compare('grade_rules_fk',$this->grade_rules_fk);
+		$criteria->compare('created_at',$this->created_at,true);
+		$criteria->compare('updated_at',$this->updated_at,true);
 
 		return new CActiveDataProvider($this, array(
 			'criteria'=>$criteria,
@@ -123,7 +114,7 @@ class GradeRules extends CActiveRecord
 	 * Returns the static model of the specified AR class.
 	 * Please note that you should have this exact method in all your CActiveRecord descendants!
 	 * @param string $className active record class name.
-	 * @return GradeRules the static model class
+	 * @return ClassroomVsGradeRules the static model class
 	 */
 	public static function model($className=__CLASS__)
 	{