+++ js/classroom/form/functions.js
@@ -62,7 +62,6 @@ var removeDiscipline = function (instructor, discipline) {
     }
     $("li[instructor = " + instructor + "] li[discipline = " + discipline + "]").remove();
 }
-
 $(document).on("change", "#Role", function () {
     $(".regent-teacher-container").hide()
     if ($(this).val() == 1 && RegentTeacherCount < 2) {