+++ themes/default/views/admin/index.php
@@ -95,7 +95,7 @@ $this->pageTitle = 'TAG - ' . Yii::t('default', 'Administration');
                     <div class="container-box">
                         <p>Unidades e avaliações</p>
 
-                        <a href="<?php echo Yii::app()->createUrl('admin/gradesStructure') ?>">
+                        <a href="<?php echo Yii::app()->createUrl('admin/indexGradesStructure') ?>">
                             <button type="button" class="admin-box-container">
                                 <div class="pull-left" style="margin-right: 20px;">
                                     <span class="t-icon-diary t-reports_icons"></span>
@@ -143,7 +143,7 @@ $this->pageTitle = 'TAG - ' . Yii::t('default', 'Administration');
                             </button>
                         </a>
                     </div>
-                
+
                     <div class="container-box">
                         <p>Configurações</p>
                         <a href="<?php echo Yii::app()->createUrl('admin/exports') ?>">